+++ rust/services/order/app/src/core/port/incoming.rs
@@ -8,7 +8,7 @@ pub type DynOrderService = Arc<dyn OrderService + Send + Sync>;
 
 #[async_trait]
 pub trait OrderService {
-    async fn get_orders(&self, offset: i32, limit: i32) -> Result<Vec<Order>, String>;
+    async fn get_orders(&self, offset: i64, limit: i64) -> Result<Vec<Order>, String>;
     async fn place_order(&self, item_names: Vec<String>) -> Result<Order, String>;
     async fn get_order(&self, order_id: OrderId) -> Result<Order, String>;
 }
